Use `SequencesTooltip` and `CollectionsTooltip` for EA cards
