+++ packages/lesswrong/lib/components.ts
@@ -865,7 +865,7 @@ importComponent("SequencesGridItem", () => require('../components/sequences/Sequ
 importComponent("LargeSequencesItem", () => require('../components/sequences/LargeSequencesItem'));
 importComponent("SequencesTooltip", () => require('../components/sequences/SequencesTooltip'));
 importComponent("SequencesSummary", () => require('../components/sequences/SequencesSummary'));
-importComponent("CollectionsHoverOver", () => require('../components/collections/CollectionsHoverOver'));
+importComponent("CollectionsTooltip", () => require('../components/collections/CollectionsTooltip'));
 importComponent("ChapterTitle", () => require('../components/sequences/ChapterTitle')); 
 importComponent("SequencesSmallPostLink", () => require('../components/sequences/SequencesSmallPostLink'));
 importComponent("ChaptersItem", () => require('../components/sequences/ChaptersItem'));
